Refactor functions to load_data
