+++ bin/data_collect/load_data.py
@@ -31,3 +31,28 @@ def get_all_folders_path(folder_path):
                     for folder in folders_path]
 
     return folders_path
+
+
+def get_csv_file_path():
+    usage_message = "Usage: {} [csv_file_path]".format(sys.argv[0])
+
+    if len(sys.argv) < 2:
+        print usage_message
+        exit(1)
+
+    csv_file_path = sys.argv[1]
+    csv_file_path = os.path.expanduser(csv_file_path)
+
+    if not os.path.exists(csv_file_path):
+        print usage_message
+        print "CSV file not exists"
+        exit(1)
+
+    return csv_file_path
+
+def get_lines_from_csv_file(csv_file_path):
+    with open(csv_file_path, 'rb') as text:
+        lines = [line.strip() for line in text]
+    lines = [line.split(';') for line in lines]
+
+    return lines
