+++ bin/data_collect/plot_user_preferences.py
@@ -6,6 +6,7 @@ import numpy as np
 import os
 import sys
 
+from load_data import get_csv_file_path, get_lines_from_csv_file
 
 def autolabel(ax, rects, string_format):
     for rect in rects:
@@ -45,7 +46,6 @@ def plot_strategies_score(strategies_score, classifications, title, ylabel,
     plt.yticks(np.arange(plot_min, plot_max, plot_step))
     plt.show()
 
-
 def get_csv_file_path():
     usage_message = "Usage: {} [csv_file_path]".format(sys.argv[0])
 
@@ -65,11 +65,7 @@ def get_csv_file_path():
 
 
 def load_csv_file(csv_file_path):
-    with open(csv_file_path, 'rb') as text:
-        lines = [line.strip() for line in text]
-
-    lines = [line.split(';') for line in lines]
-
+    lines = get_lines_from_csv_file(csv_file_path)
     scores = [[line[0]] + map(int, line[1:]) for line in lines[1:]]
     classifications = lines[0][1:]
